Refactor IHardwareMouse.h
